feat: remove util file, unfolding into src
