+++ src/constant.ts
@@ -5,3 +5,15 @@ export const DYNAMIC = 'dynamic';
 export const STATIC = 'static';
 
 export const ICESTSRK_NOT_FOUND = `/${PREFIX}_404`;
+
+export const setIcestark = (key: string, value: any) => {
+  if (!(window as any).ICESTARK) {
+    (window as any).ICESTARK = {};
+  }
+  (window as any).ICESTARK[key] = value;
+};
+
+export const getIcestark = (key: string) => {
+  const icestark: any = (window as any).ICESTARK;
+  return icestark && icestark[key] ? icestark[key] : null;
+};
